Ants move towards food, with a bit of randomness. Finally!!
Also fixed World::findNearestFood which was completely wrong.

This was so incredibly difficult its actually outrageous. I was fighting atan2 for probably about 3-5 hours, so I gave up and used a method based on the unit vector between the food and ant instead, which works a lot better. Since we're on a grid, we use Chebyshev distance. I'm really happy this worked, because it was so incredibly frustrating to get working, but it looks exactly how I wanted it to.